CM-710 Refactoring of validationSchemas
