+++ src/containers/Common/components/CommonDetailContainer/CreateProposalModal/FundsAllocationStage/Configuration/Configuration.tsx
@@ -14,7 +14,7 @@ import { Formik, FormikConfig } from "formik";
 import { FormikProps } from "formik/dist/types";
 import { Form, TextField } from "@/shared/components/Form/Formik";
 import { FUNDS_ALLOCATION_PROPOSAL_TITLE_LENGTH } from "./constants";
-import { validationSchema } from "./validationSchema";
+import { configurationValidationSchema } from "../validationSchema";
 import "./index.scss";
 
 interface ConfigurationProps {
@@ -72,7 +72,7 @@ const Configuration: FC<ConfigurationProps> = (props) => {
           initialValues={getInitialValues()}
           onSubmit={handleSubmit}
           innerRef={formRef}
-          validationSchema={validationSchema}
+          validationSchema={configurationValidationSchema}
           validateOnMount
         >
           {({ isValid }) => (
