+++ src/containers/Common/components/CommonDetailContainer/CreateProposalModal/FundsAllocationStage/FundDetails/FundDetails.tsx
@@ -23,7 +23,7 @@ import { getScreenSize } from "@/shared/store/selectors";
 import { StageName } from "../../StageName";
 import { getPrefix } from "../helpers";
 import { FundsAllocationData, FundType } from "../types";
-import { validationSchema } from "./validationSchema";
+import { fundDetailsValidationSchema } from "../validationSchema";
 import {FUND_TYPES} from '../constants';
 import "./index.scss";
 
@@ -136,7 +136,7 @@ const FundDetails: FC<ConfigurationProps> = (props) => {
           enableReinitialize
           onSubmit={handleSubmit}
           innerRef={formRef}
-          validationSchema={validationSchema}
+          validationSchema={fundDetailsValidationSchema}
           validateOnMount
         >
           {({ values, errors, touched, isValid }) => (
