--- src/containers/Common/components/CommonDetailContainer/CreateProposalModal/FundsAllocationStage/FundsAllocationForm/constants.ts
@@ -1,1 +1,0 @@
-export const FUNDS_ALLOCATION_PROPOSAL_TITLE_LENGTH = 60;
