+++ src/containers/Common/components/CommonDetailContainer/CreateProposalModal/FundsAllocationStage/validationSchema.ts
@@ -2,12 +2,15 @@ import * as yup from "yup";
 import { formatPrice } from "@/shared/utils";
 import { FUNDS_ALLOCATION_PROPOSAL_TITLE_LENGTH } from "./constants";
 
-export const validationSchema = yup.object({
+export const configurationValidationSchema = yup.object().shape({
   title: yup.string()
     .max(FUNDS_ALLOCATION_PROPOSAL_TITLE_LENGTH, "Entered title is too long")
     .required("Please enter proposal title"),
   description: yup.string().required("Please enter proposal description"),
   goalOfPayment: yup.string().required("Please enter goal of payment"),
+});
+
+export const fundDetailsValidationSchema = yup.object({
   commonBalance: yup.number(),
   amount: yup
     .number()
@@ -25,3 +28,5 @@ export const validationSchema = yup.object({
     then: yup.object().required()
   }).nullable(),
 });
+
+export const fundAllocationValidationSchema = configurationValidationSchema.concat(fundDetailsValidationSchema);
